change return type from size_t to int
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder/branches/mira-mtm5@3751 1f5c12ca-751b-0410-a591-d2e778427230

+++ mira/Perceptron.cpp
@@ -24,7 +24,7 @@ using namespace std;
 
 namespace Mira {
 
-size_t Perceptron::updateWeights(ScoreComponentCollection& currWeights,
+int Perceptron::updateWeights(ScoreComponentCollection& currWeights,
 		const vector< vector<ScoreComponentCollection> >& scores,
 		const vector<vector<float> >& losses,
 		const vector<ScoreComponentCollection>& oracleScores)
